requirements.txt
Pillow==7.1.2
pytesseract==0.3.4
responder==2.0.5
